Semantic Versioning
Ensure that people can start requiring "ircmaxell/random-lib" using
semantic versioning like the following instead of forcing people to use
`dev-master` as the version:

    "ircmaxell/random-lib": "1.*@dev"

An unintended side effect of composer.json changes is that it results in
composer.lock needing to change to account for a different hash. In addition,
Composer sometimes actually changes other aspects of the composer.lock
structure.