README++
Provide some examples and information on how to use RandomLib in the
README. Potential users ill be better able to evaluate RandomLib to
see if it will do what they need if they can see exactly what it does
and how it is accomplished.

A lot of this is copy and pasted from the docblocks but I've changed
a few bits and pieces here and there. Hopefully it maintains the
spirit of the original docs.